refactor(ui-react): extract agent result previews
